Fixed the bug in MultiCore that was yielding improper MTTFs.
Basically, at each step of the simulation, MultiCore activates the number of cores it needs to in order to make up for the failed core. The problem came in determining the number of cores that had failed, which before was determined by the total number of cores that had failed. However, this number should be the total number of cores that had failed on that particular step. It was basically a simple if check, to make sure that a care that has already failed is not processed.